Removed the change made to sleepmode when secure virtual memory is in use and changed the hibernate mode to that recommended in the OS 10.6 pmset manpage.
Various websites say you must add 4 to hibernate modes if secure vitual memory
is in use but I found on testing that this causes both machines I use to
hibernate when sleep is desired. (This must have been becuase I was adding it
to sleep mode as well, actually.) Further searching revelaed that this is no
longer recommended. It seems it was a hack to make hibernation work on old
machines. I'm not sure how old but I will leave it off and not worry unless
someone reports that they require it, then it may be possible to add something
that works out if the machine needs it.